Refactor WebP codec functions for improved memory management and error handling
- Updated `decodeRGBA` to use `std::unique_ptr` with a custom deleter for automatic memory management.
- Changed input parameter from raw pointer to `std::vector<uint8_t>` for `decodeRGBA`.
- Enhanced error messages for decoding failures.
- Modified `encodeLosslessRGBA` to accept `std::vector<uint8_t>` and improved error handling.
- Adjusted related code in `sog.cpp` to use `float` instead of `double` for codebook vectors.
- Added new GPU clustering functionality in `gpu-clustering.cu` with CUDA kernel for efficient point clustering.
- Implemented KD-Tree structure and nearest neighbor search in `kdtree.cpp`.
- Introduced K-Means clustering algorithm in `kmeans.cpp` with initialization and average calculation methods.
- Added utility functions for string manipulation in `strings.h`.
- Created new header files for GPU clustering, LCC data structures, and KD-Tree.

+++ include/splat/sog.h
@@ -43,7 +43,7 @@ struct Meta {
   } means;
 
   struct {
-    std::vector<double> codebook;
+    std::vector<float> codebook;
     std::vector<std::string> files;
   } scales;
 
@@ -52,14 +52,14 @@ struct Meta {
   } quats;
 
   struct {
-    std::vector<double> codebook;
+    std::vector<float> codebook;
     std::vector<std::string> files;
   } sh0;
 
   struct SHN {
     int count;
     int bands;
-    std::vector<double> codebook;
+    std::vector<float> codebook;
     std::vector<std::string> files;
   };
